Construct the constraint matrices for the spatial random effects alpha

Construct the constraint matrices for the spatial random effects alpha

make_constraint(MRA, constraint = "unconstrained", joint = TRUE)

Arguments

MRA

An MRA object. Usually contained in the output from mcmc_mra() or mcmc_mra_integrated() or for developers, the MRA object can be created directly from the function mra_wendland_2d()

constraint

What constraint should be applied to the spatial process? Options include no constraint (constraint = "unconstrained"), a constraint so the entire MRA process sums to 0 (constraint = "overall"), a constraint so that each of the M levels of the MRA process sum to 0 (constraint = "resolution"), or whether the predicte process must sum to 0 (constraint = "predicted"). Note: constraint = "predicted" is NOT currently supported.

joint

Should the spatial parameters alpha be sampled jointly or by each resolution

Value

A list of constraint matrices A_constraint and a_constraint for use in spam sampling functions