Bayesian Polya-gamma regression

this function runs the Bayesian multinomial regression using Polya-gamma data augmentation

mcmc_mra(
  y,
  X,
  locs,
  params,
  priors = NULL,
  M = 4,
  n_neighbors = 68,
  n_coarse_grid = 10,
  n_padding = 5L,
  n_cores = 1L,
  inits = NULL,
  config = NULL,
  verbose = FALSE,
  use_spam = TRUE,
  n_chain = 1
)

Arguments

y
is a \(n\) vector of Gaussian data.

X
is a \(n \times p\) matrix of fixed effects (like latitude, elevation, etc)

locs
is a \(n \times 2\) matrix of observation locations.

params
is a list of parameter settings. The list params must contain the following values:

n_adapt: A positive integer number of adaptive MCMC iterations.

n_mcmc: A positive integer number of total MCMC iterations post adaptation.

n_thin: A positive integer number of MCMC iterations per saved sample.

n_message: A positive integer number of frequency of iterations to output a progress message. For example, n_message = 50 outputs progress messages every 50 iterations.

priors
is the list of prior settings.

M
The number of resolutions.

n_neighbors
The expected number of neighbors for each interior basis function. This determines the basis radius parameter.

n_coarse_grid
The number of basis functions in one direction (e.g. n_coarse_grid = 10 results in a \(10 \times 10\) course grid which is further extended by the number of additional padding basis functions given by n_padding.

n_padding
The number of additional boundary points to add on each boundary. For example, n_padding = 5 will add 5 boundary knots to the both the left and right side of the grid).

n_cores
is the number of cores for parallel computation using openMP.

inits
is the list of initial values if the user wishes to specify initial values. If these values are not specified, then the initial values will be randomly sampled from the prior.

config
is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.

verbose
Should verbose output be printed? Typically this is only useful for troubleshooting.

use_spam
is a boolean flag to determine whether the output is a list of spam matrix objects (use_spam = TRUE) or a an \(n \times n\) sparse Matrix of class "dgCMatrix" use_spam = FALSE (see spam and Matrix packages for details).

n_chain
is the MCMC chain id. The default is 1.