this function updates a block Gaussian random walk proposal Metropolis-Hastings tuning parameters

this function updates a block Gaussian random walk proposal Metropolis-Hastings tuning parameters

update_tuning_mv(k, accept, lambda, batch_samples, Sigma_tune, Sigma_tune_chol)

Arguments

k
is a positive integer that is the current MCMC iteration.

accept
is a positive number between 0 and 1 that represents the batch accpetance rate. The target of this adaptive tuning algorithm is an acceptance rate of between 0.44 (a univariate update) and 0.234 (5+ dimensional upate).

lambda
is a positive scalar that scales the covariance matrix Sigma_tune and diminishes towards 0 as k increases.

batch_samples
is a \(50 \times d\) dimensional matrix that consists of the 50 batch samples of the d-dimensional parameter being sampled.

Sigma_tune
is a \(d \times d\) positive definite covariance matrix of the batch samples used to generate the multivariate Gaussian random walk proposal.

Sigma_tune_chol
is the \(d \times d\) Cholesky decomposition of Sigma_tune.